Move Auth entites to authentication lib
