+++ apps/modernization-api/src/main/java/gov/cdc/nbs/patient/PatientService.java
@@ -35,6 +35,7 @@ import org.springframework.stereotype.Service;
 import com.blazebit.persistence.CriteriaBuilderFactory;
 import com.blazebit.persistence.querydsl.BlazeJPAQuery;
 import com.querydsl.core.types.dsl.BooleanExpression;
+import gov.cdc.nbs.authentication.UserService;
 import gov.cdc.nbs.config.security.SecurityUtil;
 import gov.cdc.nbs.entity.elasticsearch.ElasticsearchPerson;
 import gov.cdc.nbs.entity.enums.RecordStatus;
@@ -57,7 +58,6 @@ import gov.cdc.nbs.model.PatientEventResponse;
 import gov.cdc.nbs.patient.create.PatientCreateRequestResolver;
 import gov.cdc.nbs.patient.kafka.KafkaProducer;
 import gov.cdc.nbs.repository.PersonRepository;
-import gov.cdc.nbs.service.UserService;
 import gov.cdc.nbs.time.FlexibleInstantConverter;
 import graphql.com.google.common.collect.Ordering;
 import lombok.RequiredArgsConstructor;
