+++ apps/modernization-api/src/main/java/gov/cdc/nbs/redirect/incoming/ClassicIncomingContextResolver.java
@@ -1,6 +1,6 @@
 package gov.cdc.nbs.redirect.incoming;
 
-import gov.cdc.nbs.AuthorizedUserResolver;
+import gov.cdc.nbs.authentication.AuthorizedUserResolver;
 import gov.cdc.nbs.authorization.NBSUserCookie;
 import gov.cdc.nbs.authorization.SessionCookie;
 import org.springframework.stereotype.Component;
