+++ apps/modernization-api/src/test/java/gov/cdc/nbs/TestContext.java
@@ -1,7 +1,7 @@
 package gov.cdc.nbs;
 
 import org.springframework.test.web.servlet.MvcResult;
-import gov.cdc.nbs.authentication.AuthUser;
+import gov.cdc.nbs.authentication.entity.AuthUser;
 
 /**
  * Used to store data that will be shared between steps. Always call TestContext.clear() in a @Before block
