+++ apps/modernization-api/src/test/java/gov/cdc/nbs/patient/profile/contact/PatientProfileViewContactSteps.java
@@ -2,7 +2,6 @@ package gov.cdc.nbs.patient.profile.contact;
 
 import gov.cdc.nbs.authorization.SessionCookie;
 import gov.cdc.nbs.patient.TestPatientIdentifier;
-import gov.cdc.nbs.patient.TestPatients;
 import gov.cdc.nbs.patient.contact.TestContactTracings;
 import gov.cdc.nbs.support.TestActive;
 import io.cucumber.java.Before;
