+++ apps/modernization-api/src/test/java/gov/cdc/nbs/support/util/UserUtil.java
@@ -1,8 +1,8 @@
 package gov.cdc.nbs.support.util;
 
 import org.springframework.security.core.context.SecurityContextHolder;
-import gov.cdc.nbs.authentication.AuthUser;
-import gov.cdc.nbs.authentication.AuthUserRepository;
+import gov.cdc.nbs.authentication.entity.AuthUser;
+import gov.cdc.nbs.authentication.entity.AuthUserRepository;
 import gov.cdc.nbs.config.security.NbsUserDetails;
 
 public class UserUtil {
