+++ apps/patient-listener/src/main/java/gov/cdc/nbs/patientlistener/request/create/PatientCreateRequestHandler.java
@@ -1,12 +1,12 @@
 package gov.cdc.nbs.patientlistener.request.create;
 
+import gov.cdc.nbs.authentication.UserService;
 import gov.cdc.nbs.entity.odse.Person;
 import gov.cdc.nbs.message.patient.event.PatientCreateData;
 import gov.cdc.nbs.patientlistener.request.UserNotAuthorizedException;
 import gov.cdc.nbs.patientlistener.request.PatientRequestStatusProducer;
 import gov.cdc.nbs.patientlistener.util.PersonConverter;
 import gov.cdc.nbs.repository.elasticsearch.ElasticsearchPersonRepository;
-import gov.cdc.nbs.service.UserService;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -15,57 +15,57 @@ import org.springframework.transaction.annotation.Transactional;
 @Slf4j
 @Service
 public class PatientCreateRequestHandler {
-  @Autowired
-  private UserService userService;
+    @Autowired
+    private UserService userService;
 
-  @Autowired
-  private ElasticsearchPersonRepository elasticsearchPersonRepository;
+    @Autowired
+    private ElasticsearchPersonRepository elasticsearchPersonRepository;
 
-  @Autowired
-  private PatientRequestStatusProducer statusProducer;
+    @Autowired
+    private PatientRequestStatusProducer statusProducer;
 
-  @Autowired
-  PatientCreator creator;
+    @Autowired
+    PatientCreator creator;
 
-  @Transactional
-  public void handlePatientCreate(PatientCreateData data) {
-    long user = data.createdBy();
-    log.debug("Attempting to validate permissions for user: {}", user);
-    if (userService.isAuthorized(user, "FIND-PATIENT", "ADD-PATIENT")) {
-      creationAllowed(data.request(), data);
-    } else {
-      notAuthorized(data.request());
+    @Transactional
+    public void handlePatientCreate(PatientCreateData data) {
+        long user = data.createdBy();
+        log.debug("Attempting to validate permissions for user: {}", user);
+        if (userService.isAuthorized(user, "FIND-PATIENT", "ADD-PATIENT")) {
+            creationAllowed(data.request(), data);
+        } else {
+            notAuthorized(data.request());
+        }
     }
-  }
 
-  private void creationAllowed(final String key, final PatientCreateData createRequest) {
-    // perform the creation
-    log.debug("User permission validated. Creating patient");
-    Person newPatient = createPatient(createRequest);
-    createElasticsearchPatient(newPatient);
+    private void creationAllowed(final String key, final PatientCreateData createRequest) {
+        // perform the creation
+        log.debug("User permission validated. Creating patient");
+        Person newPatient = createPatient(createRequest);
+        createElasticsearchPatient(newPatient);
 
-    // post success message to status topic
-    statusProducer.successful(key, "Successfully created patient", newPatient.getId());
-  }
+        // post success message to status topic
+        statusProducer.successful(key, "Successfully created patient", newPatient.getId());
+    }
 
-  private void notAuthorized(final String key) {
-    log.debug("User lacks permission for patient create");
-    throw new UserNotAuthorizedException(key);
-  }
+    private void notAuthorized(final String key) {
+        log.debug("User lacks permission for patient create");
+        throw new UserNotAuthorizedException(key);
+    }
 
-  /**
-   * Creates a Person entity from the PatientInput and persists it to the database
-   */
-  private Person createPatient(PatientCreateData request) {
-    return creator.create(request);
-  }
+    /**
+     * Creates a Person entity from the PatientInput and persists it to the database
+     */
+    private Person createPatient(PatientCreateData request) {
+        return creator.create(request);
+    }
 
-  /**
-   * Converts a Person entity into an ElasticsearchPerson Document and persists it to ES
-   */
-  private void createElasticsearchPatient(Person person) {
-    var esPerson = PersonConverter.toElasticsearchPerson(person);
-    elasticsearchPersonRepository.save(esPerson);
-  }
+    /**
+     * Converts a Person entity into an ElasticsearchPerson Document and persists it to ES
+     */
+    private void createElasticsearchPatient(Person person) {
+        var esPerson = PersonConverter.toElasticsearchPerson(person);
+        elasticsearchPersonRepository.save(esPerson);
+    }
 
 }
