+++ apps/question-bank/src/main/java/gov/cdc/nbs/questionbank/config/security/JWTFilter.java
@@ -16,9 +16,9 @@ import org.springframework.web.filter.OncePerRequestFilter;
 import com.auth0.jwt.JWTVerifier;
 import com.auth0.jwt.exceptions.JWTVerificationException;
 import com.auth0.jwt.interfaces.DecodedJWT;
+import gov.cdc.nbs.authentication.UserService;
 import gov.cdc.nbs.config.security.NbsUserDetails;
 import gov.cdc.nbs.config.security.SecurityProperties;
-import gov.cdc.nbs.service.UserService;
 import lombok.RequiredArgsConstructor;
 
 @Component
