+++ libs/authentication/build.gradle
@@ -29,8 +29,6 @@ dependencyManagement {
     }
 }
 dependencies {
-    implementation project(':database-entities')
-
     // core
     implementation libs.bundles.jpa
     annotationProcessor libs.springBoot.jpa
@@ -52,6 +50,8 @@ dependencies {
     compileOnly libs.lombok
     annotationProcessor libs.lombok
 
-    // Kafka
-    implementation libs.spring.kafka
+}
+
+compileJava {
+    options.compilerArgs << '-Aquerydsl.generatedAnnotationClass=com.querydsl.core.annotations.Generated'
 }
