+++ libs/authentication/src/main/java/gov/cdc/nbs/authentication/entity/AuthPermSetRepository.java
@@ -1,4 +1,4 @@
-package gov.cdc.nbs.authentication;
+package gov.cdc.nbs.authentication.entity;
 
 import org.springframework.data.jpa.repository.JpaRepository;
