+++ libs/authentication/src/main/java/gov/cdc/nbs/authentication/enums/RecordStatus.java
@@ -1,0 +1,8 @@
+package gov.cdc.nbs.authentication.enums;
+
+public enum RecordStatus {
+    ACTIVE,
+    INACTIVE,
+    LOG_DEL,
+    SUPERCEDED
+}
