+++ libs/authentication/src/main/java/gov/cdc/nbs/authentication/repository/SecurityLogRepository.java
@@ -1,12 +1,12 @@
-package gov.cdc.nbs.repository;
+package gov.cdc.nbs.authentication.repository;
 
 import java.util.List;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.querydsl.QuerydslPredicateExecutor;
+import gov.cdc.nbs.authentication.entity.SecurityLog;
 
-import gov.cdc.nbs.entity.odse.SecurityLog;
 
 public interface SecurityLogRepository
         extends JpaRepository<SecurityLog, Long>, QuerydslPredicateExecutor<SecurityLog> {
